test(rstest): migrate chunk-loading config cases
