+++ packages/webpack/chunk-loading-webpack-plugin/package.json
@@ -40,10 +40,9 @@
     "@lynx-js/webpack-runtime-globals": "workspace:*"
   },
   "devDependencies": {
-    "@lynx-js/test-tools": "workspace:*",
     "@microsoft/api-extractor": "catalog:",
     "@rspack/core": "catalog:rspack",
-    "@rspack/test-tools": "catalog:rspack",
+    "@rspack/test-tools": "1.7.9",
     "@rstest/core": "catalog:rstest",
     "css-loader": "^7.1.4",
     "mini-css-extract-plugin": "^2.10.0",
