--- packages/webpack/chunk-loading-webpack-plugin/test/cases/chunk-loading/development/dynamic.js
@@ -1,1 +1,0 @@
-export const value = 1;
