--- packages/webpack/chunk-loading-webpack-plugin/test/cases/chunk-loading/production/webpack.config.js
@@ -1,17 +1,0 @@
-import { ChunkLoadingWebpackPlugin } from '../../../../lib/index.js';
-
-/** @type {import('webpack').Configuration} */
-export default {
-  mode: 'production',
-  output: {
-    chunkLoading: 'lynx',
-    chunkFormat: 'commonjs',
-    chunkFilename: '[id].webpack.bundle.cjs',
-  },
-  optimization: {
-    chunkIds: 'named',
-  },
-  plugins: [
-    new ChunkLoadingWebpackPlugin(),
-  ],
-};
