--- packages/webpack/chunk-loading-webpack-plugin/test/cases/chunk-loading/runtime-globals/dynamic.js
@@ -1,1 +1,0 @@
-export const value = 1;
