--- packages/webpack/chunk-loading-webpack-plugin/test/cases/css/initial/index.js
@@ -1,11 +1,0 @@
-/// <reference types="@rspack/test-tools/rstest" />
-
-import './foo.css';
-
-it('should install the CSS chunk', async function() {
-  const hasChunkInstalled = Object.keys(__webpack_require__.O).every((
-    key,
-  ) => (__webpack_require__.O[key]('css-css_initial_foo_css')));
-
-  expect(hasChunkInstalled).toBe(true);
-});
