+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/chunk-loading/development/rspack.config.js
@@ -1,0 +1,20 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+import { ChunkLoadingWebpackPlugin } from '../../../../lib/index.js';
+
+/** @type {import('@rspack/core').Configuration} */
+export default {
+  mode: 'development',
+  output: {
+    chunkLoading: 'lynx',
+    chunkFormat: 'commonjs',
+    chunkFilename: '[id].rspack.bundle.cjs',
+  },
+  plugins: [
+    new ChunkLoadingWebpackPlugin(),
+    compiler => {
+      new compiler.webpack.HotModuleReplacementPlugin().apply(compiler);
+    },
+  ],
+};
