+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/chunk-loading/production/dynamic.js
@@ -1,0 +1,4 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+export const value = 1;
