+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/chunk-loading/production/webpack.config.js
@@ -1,12 +1,15 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
 import { ChunkLoadingWebpackPlugin } from '../../../../lib/index.js';
 
-/** @type {import('@rspack/core').Configuration} */
+/** @type {import('webpack').Configuration} */
 export default {
   mode: 'production',
   output: {
     chunkLoading: 'lynx',
     chunkFormat: 'commonjs',
-    chunkFilename: '[id].rspack.bundle.cjs',
+    chunkFilename: '[id].webpack.bundle.cjs',
   },
   optimization: {
     chunkIds: 'named',
