+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/chunk-loading/runtime-globals/rspack.config.js
@@ -1,3 +1,6 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
 import { RuntimeGlobals } from '@lynx-js/webpack-runtime-globals';
 
 import { ChunkLoadingWebpackPlugin } from '../../../../lib/index.js';
