+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/css/development/index.js
@@ -1,11 +1,16 @@
+// Copyright 2024 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+
 /// <reference types="@rspack/test-tools/rstest" />
 
 import fs from 'node:fs';
 import { createRequire } from 'node:module';
 import path from 'node:path';
 
-const require = createRequire(__filename);
+/* global __filename, __dirname, rstest */
 
+const require = createRequire(__filename);
 globalThis.lynx = {
   requireModuleAsync: rstest.fn(function requireModuleAsync(request, callback) {
     return Promise.resolve().then(() => {
@@ -19,21 +24,15 @@ globalThis.lynx = {
 };
 
 it('should have dynamic chunks', () => {
+  const assets = fs.readdirSync(__dirname);
+
   expect(
-    fs.existsSync(
-      path.join(
-        __dirname,
-        'chunk-loading_production_dynamic_js.'
-          + path.basename(__filename).replace('.js', '.cjs'),
-      ),
-    ),
+    assets.some(asset => asset.endsWith('_dynamic_css.rspack.bundle.css')),
   ).toBeTruthy();
 });
 
 it('should work with chunk loading require', async function() {
-  await import('./dynamic.js').then(module => {
-    expect(module.value).toBe(1);
-    expect(lynx.requireModuleAsync).toBeCalled();
+  await import('./dynamic.css').then(module => {
+    expect(module).not.toBeUndefined();
   });
-  expect.assertions(3);
 });
