+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/css/development/rspack.config.js
@@ -1,4 +1,8 @@
-import rspack from '@rspack/core';
+// Copyright 2024 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+
+import { rspack } from '@rspack/core';
 
 import { ChunkLoadingWebpackPlugin } from '../../../../lib/index.js';
