+++ packages/webpack/chunk-loading-webpack-plugin/test/configCases/css/initial/foo.css
